Simplify header logic to abstract from submenu names
