+++ blocks/header/header-megamenu.js
@@ -1,41 +1,52 @@
 import handleViewportChanges from './header-events.js';
 import { fetchHeaderContent } from './header.js';
-import buildProductsMegaMenu from './menus/products.js';
-import buildMegaMenuLeftMenus from './menus/left-submenus.js';
-import { getMetadata, toClassName } from '../../scripts/lib-franklin.js';
+import buildRightSubmenu from './header-megamenu-components.js';
+import { getMetadata } from '../../scripts/lib-franklin.js';
 
 function buildMegaMenu(navContent, submenuContent) {
   const productsSubmenu = document.createElement('div');
-  productsSubmenu.append(submenuContent.querySelector('h1'));
-  productsSubmenu.append(buildMegaMenuLeftMenus(submenuContent));
+  const title = submenuContent.querySelector('h1');
+  productsSubmenu.append(title.cloneNode(true));
+
+  // get div after h1
+  const divAfterH1 = submenuContent.querySelector('h1').nextElementSibling;
+  productsSubmenu.append(buildRightSubmenu(divAfterH1));
+
+  // get all H2s and create a list of them
+  const h2s = [...submenuContent.querySelectorAll('h2')];
+  const h2List = document.createElement('ul');
+  h2List.classList.add('menu-nav-submenu-sections');
+
+  // add H2s to list
+  h2s.forEach((h2) => {
+    const h2ListItem = document.createElement('li');
+    h2ListItem.classList.add('menu-nav-submenu-section');
+    h2ListItem.innerHTML = h2.outerHTML;
+
+    h2ListItem.append(buildRightSubmenu(h2));
+
+    h2List.append(h2ListItem);
+  });
+
+  productsSubmenu.append(h2List);
+
   submenuContent.innerHTML = productsSubmenu.outerHTML;
   const backgroundImg = navContent.querySelector('.submenu-background img');
   submenuContent.style.backgroundImage = `url(${backgroundImg.src})`;
 }
 
-function createSubmenuBuildersMap() {
-  // create map of submenu name to function
-  const submenus = new Map();
-  submenus.set('products', buildProductsMegaMenu);
-  submenus.set('applications', buildMegaMenu);
-  submenus.set('resources', buildMegaMenu);
-  submenus.set('service-support', buildMegaMenu);
-  submenus.set('company', buildMegaMenu);
-  submenus.set('contact-us', () => { });
-  return submenus;
+function getSubmenus() {
+  return ['products', 'applications', 'resources', 'service-support', 'company', 'contact-us'];
 }
 
 export default async function fetchAndStyleMegamenus(headerBlock) {
   // ------ Submenus ------
-  const submenuBuildersMap = createSubmenuBuildersMap();
-
-  // get all keys from submenuBuildersMap
-  const submenuKeys = [...submenuBuildersMap.keys()];
+  const submenusList = getSubmenus();
 
   // Fetch all submenu content concurrently
   const submenuFetchPromises = [];
-  for (let i = 0; i < submenuKeys.length - 1; i += 1) {
-    const submenuId = submenuKeys[i];
+  for (let i = 0; i < submenusList.length - 1; i += 1) {
+    const submenuId = submenusList[i];
     const submenuPath = getMetadata(`${submenuId}-submenu`) || `/fragments/menu/${submenuId}`;
     submenuFetchPromises.push(
       fetch(`${submenuPath}.plain.html`, window.location.pathname.endsWith(`/${submenuId}`) ? { cache: 'reload' } : {}),
@@ -55,7 +66,7 @@ export default async function fetchAndStyleMegamenus(headerBlock) {
       const closeButton = document.createElement('div');
       closeButton.classList.add('menu-nav-submenu-close');
 
-      const submenuId = submenuKeys[i];
+      const submenuId = submenusList[i];
       // eslint-disable-next-line no-await-in-loop
       const submenuHtml = await submenuResponse.text();
       const submenuContent = document.createElement('div');
@@ -63,8 +74,7 @@ export default async function fetchAndStyleMegamenus(headerBlock) {
       submenuContent.innerHTML = submenuHtml;
 
       // Get submenu builder, and build submenu
-      const submenuBuilder = submenuBuildersMap.get(toClassName(submenuId));
-      submenuBuilder(headerContent, submenuContent);
+      buildMegaMenu(headerContent, submenuContent);
 
       // Get the list item in the header block that contains a div with attribute menu-id
       // that matches the submenuId
